+++ ChangeLog.md
@@ -1,7 +1,7 @@
 # Change Log
 
 ### 1.4.3 - 2020/04/04
-- Changed the internal directory structure to load classes properly.
+- Fixed an issue that some classes were not loaded when the project is imported via Composer.
 
 ### 1.4.2 - 2020/04/03
 - Changed the internal directory structure to support Composer class autoload.